+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Components/Addresses/Rules/StreetCodeLengthRule.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.Addresses.Rules
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components.Addresses.Rules
 {
     public class StreetCodeLengthRule : IBusinessRule
     {
